starting to suspect I suffer from some (hopefully low key) form of mental retardation. Swear to god I googled left and right and then was wondering why I keep seeing value not found. I wasn't giving the lastSelectedCity a value at the start, that's why. Morale beaten shot and drowned

+++ app/src/main/java/com/example/android/fragmentpasswithviewmodel/model/day/Wind.java
@@ -11,7 +11,7 @@ public class Wind {
     private Float speed;
     @SerializedName("deg")
     @Expose
-    private Integer deg;
+    private Float deg;
 
     /**
      * No args constructor for use in serialization
@@ -25,7 +25,7 @@ public class Wind {
      * @param speed
      * @param deg
      */
-    public Wind(Float speed, Integer deg) {
+    public Wind(Float speed, Float deg) {
         super();
         this.speed = speed;
         this.deg = deg;
@@ -48,11 +48,11 @@ public class Wind {
         return Math.round(deg);
     }
 
-    public void setDeg(Integer deg) {
+    public void setDeg(Float deg) {
         this.deg = deg;
     }
 
-    public Wind withDeg(Integer deg) {
+    public Wind withDeg(Float deg) {
         this.deg = deg;
         return this;
     }
